refactor(mls): split MlsService trait into service/api.rs
